+++ src/lib.rs
@@ -1,0 +1,4 @@
+#![deny(clippy::all)]
+#![deny(clippy::pedantic)]
+
+pub mod schema;